@@ -8,9 +8,10 @@ import com.luixtech.utilities.exception.DataNotFoundException;
 import com.luixtech.utilities.exception.DuplicationException;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
+import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springdoc.api.annotations.ParameterObject;
+import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpHeaders;
@@ -18,7 +19,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
 import java.util.List;
 
 import static com.luixtech.springbootframework.utils.HttpHeaderUtils.generatePageHeaders;
@@ -57,16 +57,14 @@ public class AuthorityController {
 
     @Operation(summary = "find authority by name")
     @GetMapping("/api/authorities/{name}")
-    public ResponseEntity<Authority> findById(
-            @Parameter(description = "authority name", required = true) @PathVariable String name) {
+    public ResponseEntity<Authority> findById(@Parameter(description = "authority name", required = true) @PathVariable String name) {
         Authority domain = authorityRepository.findById(name).orElseThrow(() -> new DataNotFoundException(name));
         return ResponseEntity.ok(domain);
     }
 
     @Operation(summary = "update authority")
     @PutMapping("/api/authorities")
-    public ResponseEntity<Void> update(
-            @Parameter(description = "new authority", required = true) @Valid @RequestBody Authority domain) {
+    public ResponseEntity<Void> update(@Parameter(description = "new authority", required = true) @Valid @RequestBody Authority domain) {
         log.debug("REST request to update authority: {}", domain);
         authorityRepository.findById(domain.getName()).orElseThrow(() -> new DataNotFoundException(domain.getName()));
         authorityRepository.save(domain);
