@@ -1,22 +1,21 @@
 package com.luixtech.rpc.webcenter.controller;
 
-import com.codahale.metrics.annotation.Timed;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.luixtech.springbootframework.component.HttpHeaderCreator;
 import com.luixtech.rpc.webcenter.domain.RpcScheduledTask;
 import com.luixtech.rpc.webcenter.exception.DataNotFoundException;
 import com.luixtech.rpc.webcenter.repository.RpcScheduledTaskRepository;
 import com.luixtech.rpc.webcenter.service.RpcScheduledTaskService;
 import com.luixtech.rpc.webcenter.utils.HttpHeaderUtils;
+import com.luixtech.springbootframework.component.HttpHeaderCreator;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.Validate;
-import org.springdoc.api.annotations.ParameterObject;
+import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
@@ -24,12 +23,10 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.scheduling.config.CronTask;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 
-import static com.luixtech.springbootframework.config.apidoc.SpringDocConfiguration.AUTH;
 import static com.luixtech.rpc.core.constant.ConsumerConstants.FAULT_TOLERANCE_VAL_FAILFAST;
 import static com.luixtech.rpc.core.constant.ConsumerConstants.FAULT_TOLERANCE_VAL_FAILOVER;
 import static com.luixtech.rpc.webcenter.LuixRpcWebCenterApplication.DEFAULT_REG;
@@ -39,7 +36,6 @@ import static com.luixtech.rpc.webcenter.LuixRpcWebCenterApplication.DEFAULT_REG
  * REST controller for managing RPC scheduled tasks.
  */
 @RestController
-@SecurityRequirement(name = AUTH)
 @AllArgsConstructor
 @Slf4j
 public class RpcScheduledTaskController {
@@ -49,7 +45,6 @@ public class RpcScheduledTaskController {
 
     @Operation(summary = "create scheduled task")
     @PostMapping("/api/rpc-scheduled-tasks")
-    @Timed
     public ResponseEntity<Void> create(@Parameter(description = "task", required = true) @Valid @RequestBody RpcScheduledTask domain) {
         log.debug("REST request to create scheduled task: {}", domain);
         if (domain.getStartTime() != null && domain.getStopTime() != null) {
@@ -78,7 +73,6 @@ public class RpcScheduledTaskController {
 
     @Operation(summary = "find scheduled task list")
     @GetMapping("/api/rpc-scheduled-tasks")
-    @Timed
     public ResponseEntity<List<RpcScheduledTask>> find(@ParameterObject Pageable pageable,
                                                        @Parameter(description = "registry url identity", required = true, schema = @Schema(defaultValue = DEFAULT_REG)) @RequestParam(value = "registryIdentity") String registryIdentity,
                                                        @Parameter(description = "Task name(fuzzy query)") @RequestParam(value = "name", required = false) String name,
@@ -93,7 +87,6 @@ public class RpcScheduledTaskController {
 
     @Operation(summary = "find scheduled task by id")
     @GetMapping("/api/rpc-scheduled-tasks/{id}")
-    @Timed
     public ResponseEntity<RpcScheduledTask> findById(@Parameter(description = "task ID", required = true) @PathVariable String id) {
         RpcScheduledTask task = rpcScheduledTaskRepository.findById(id).orElseThrow(() -> new DataNotFoundException(id));
         return ResponseEntity.ok(task);
@@ -101,7 +94,6 @@ public class RpcScheduledTaskController {
 
     @Operation(summary = "update scheduled task")
     @PutMapping("/api/rpc-scheduled-tasks")
-    @Timed
     public ResponseEntity<Void> update(@Parameter(description = "new task", required = true) @Valid @RequestBody RpcScheduledTask domain) {
         log.debug("REST request to update scheduled task: {}", domain);
         if (domain.getStartTime() != null && domain.getStopTime() != null) {
@@ -128,7 +120,6 @@ public class RpcScheduledTaskController {
 
     @Operation(summary = "delete scheduled task by id", description = "The data may be referenced by other data, and some problems may occur after deletion")
     @DeleteMapping("/api/rpc-scheduled-tasks/{id}")
-    @Timed
     public ResponseEntity<Void> delete(@Parameter(description = "task ID", required = true) @PathVariable String id) {
         log.debug("REST request to delete scheduled task: {}", id);
         rpcScheduledTaskService.delete(id);
