@@ -1,16 +1,11 @@
 package com.luixtech.rpc.webcenter.controller;
 
-import com.luixtech.springbootframework.config.LuixProperties;
+import com.luixtech.rpc.webcenter.config.dbmigrations.InitialSetupMigration;
 import com.luixtech.rpc.webcenter.domain.Authority;
+import com.luixtech.springbootframework.config.LuixProperties;
 import com.luixtech.utilities.network.AddressUtils;
-import io.mongock.api.config.MongockConfiguration;
-import io.mongock.driver.api.driver.ConnectionDriver;
-import io.mongock.driver.mongodb.springdata.v3.config.MongoDBConfiguration;
-import io.mongock.driver.mongodb.springdata.v3.config.SpringDataMongoV3Context;
-import io.mongock.runner.springboot.MongockSpringboot;
-import io.mongock.runner.springboot.RunnerSpringbootBuilder;
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import jakarta.annotation.Resource;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
@@ -27,16 +22,12 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.annotation.Resource;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
-import static com.luixtech.springbootframework.config.apidoc.SpringDocConfiguration.AUTH;
-
 
 @RestController
-@SecurityRequirement(name = AUTH)
 @Slf4j
 public class SystemController {
 
@@ -61,9 +52,7 @@ public class SystemController {
     @Resource
     private ApplicationEventPublisher            applicationEventPublisher;
     @Resource
-    private MongockConfiguration                 mongockConfiguration;
-    @Resource
-    private MongoDBConfiguration                 mongoDBConfiguration;
+    private InitialSetupMigration                initialSetupMigration;
     @Resource
     private Optional<PlatformTransactionManager> txManagerOpt;
 
@@ -116,16 +105,9 @@ public class SystemController {
 
     @Operation(summary = "reset database")
     @GetMapping("/open-api/systems/reset-database")
-    public String resetDatabase() {
-        mongoTemplate.getDb().drop();
-        ConnectionDriver connectionDriver = new SpringDataMongoV3Context()
-                .connectionDriver(mongoTemplate, mongockConfiguration, mongoDBConfiguration, txManagerOpt);
-        RunnerSpringbootBuilder runnerSpringbootBuilder = MongockSpringboot.builder()
-                .setDriver(connectionDriver)
-                .setConfig(mongockConfiguration)
-                .setSpringContext(applicationContext)
-                .setEventPublisher(applicationEventPublisher);
-        runnerSpringbootBuilder.buildRunner().execute();
+    public String resetDatabase() throws Exception {
+        initialSetupMigration.drop();
+        initialSetupMigration.run(null);
         return "Reset database successfully.";
     }
 }
