@@ -1,9 +1,9 @@
 package com.luixtech.rpc.webcenter.dto;
 
+import jakarta.validation.constraints.NotEmpty;
 import lombok.Builder;
 import lombok.Data;
 
-import javax.validation.constraints.NotEmpty;
 
 @Data
 @Builder
